General improvements to get_supervised_procs
The core issues this patch attempts to solve is two fold, 1) have
release_handler_1 act slightly differently in two corner cases and 2)
clean up the code in get_supervised_procs/0 to remove nested cases and
etc.

Regarding #1, get_supervised_procs/0 will now call functions to
test to see if the supervisor is suspended before attempting to ask it
for a list of children. This prevents a timeout and VM restart *before*
installing the release, thus never upgrading. It now will return an
empty list and print an error message allowing the upgrade to proceed.

The second corner case is if in a child specification a supervisor is
incorrectly stated to be a worker and get_modules is called against it.
Again a timeout occurs causing a timeout and VM restart before
installing the release. Similar to the last corner case in this patch
a empty list is returned and a error message is printed to allow the
upgrade to continue.

When first looking into the issue it was hard to discover why my
upgrades where failing. All I received during the upgrade process was
a timeout and a VM restart, no other information. So even if the
behavior I described above is unwanted I believe a patch should
be included with more descriptive output to help users track down
issues like these.

Regarding #2, due to the above confusion in trying to figure out what
had happened I dug into the code and started tracing it through and
found that the nested case statements and etc made it confusing. So I
started to rework and clean up, hopefully making this code path clearer
to future readers.