Fix incorrect use of error_logger:(info|warning|error)_report/2
Change incorrect calls to the report/2 functions that are supplying a format
string and a list of terms to print to use the correct
(info|warning|error)_msg/2 functions instead. The incorrect use of these
functions is generating a report with the type set to the format string, which
is ignored by both the builtin error_logger handlers and the SASL ones.